Add structs file, change variable name, delete unused code
