+++ goyaml/goyaml.go
@@ -29,22 +29,17 @@ func GoYAMLGeneration(packageSource string, ymlOutput string, packagePrefix stri
 
 	// initialization
 	ns := vfs.NameSpace{}
-	ns.Bind("/", vfs.OS("C:/Go"), "/", vfs.BindReplace)
 	ns.Bind("/src", vfs.OS(packageSource), "/", vfs.BindAfter)
-	c := godoc.NewCorpus(ns)
-	p := godoc.NewPresentation(c)
+	corpus := godoc.NewCorpus(ns)
+	pres := godoc.NewPresentation(corpus)
 
 	// Begin of get package info
-	abspath, relpath := paths(ns, p, packageName)
+	abspath, relpath := paths(ns, pres, packageName)
 	var mode godoc.PageInfoMode
 	mode = godoc.NoHTML
 	var info *godoc.PageInfo
-	info = p.GetPkgPageInfo(abspath, relpath, mode)
+	info = pres.GetPkgPageInfo(abspath, relpath, mode)
 	// end of get package info
-
-	// Begin: get position
-	// PrintPosition(info) - only for test purpose
-	// End: get position
 
 	// to YAML struct
 	docPackage, types := ToDocfx(info, packageName, packagePrefix)
@@ -79,7 +74,7 @@ func GoYAMLGeneration(packageSource string, ymlOutput string, packagePrefix stri
 }
 
 func PrintYaml(yamlBytes []byte, outputPath string, fileName string) error {
-	yamlFile, err := os.Create(outputPath + "/" + fileName + ".yml")
+	yamlFile, err := os.Create(fmt.Sprintf("%s/%s.yml", outputPath, fileName))
 	if err != nil {
 		fmt.Errorf("Failed to create file: ", fileName)
 		return err
@@ -89,47 +84,6 @@ func PrintYaml(yamlBytes []byte, outputPath string, fileName string) error {
 	yamlFile.Write(yamlBytes)
 	yamlFile.Close()
 	return nil
-}
-
-// PrintPosition
-// some Sample codes to print the source code position, for constants & functions & methods
-func PrintPosition(info *godoc.PageInfo) {
-	// print position info for constant
-	fmt.Println("---Constant source info example--------------------")
-	if len(info.PDoc.Consts) > 0 {
-		constant := info.PDoc.Consts[0] // use the first one for example
-		position := constant.Decl.Pos()
-		fs := info.FSet.Position(position)
-		fmt.Println(constant.Names)
-		fmt.Println(fs.Filename, fs.Line)
-	}
-	fmt.Println("---Function source info example--------------------")
-	// print position info for function
-	if len(info.PDoc.Funcs) > 0 {
-		function := info.PDoc.Funcs[0]
-		position := function.Decl.Pos()
-		fs := info.FSet.Position(position)
-		fmt.Println(function.Name)
-		fmt.Println(fs.Filename, fs.Line)
-	}
-	fmt.Println("---Type source info example--------------------")
-	// print position info for type
-	if len(info.PDoc.Types) > 0 {
-		typea := info.PDoc.Types[0]
-		position := typea.Decl.Pos()
-		fs := info.FSet.Position(position)
-		fmt.Println(typea.Name)
-		fmt.Println(fs.Filename, fs.Line)
-		if len(typea.Methods) > 0 {
-			fmt.Println("---Method source info example--------------------")
-			methoda := typea.Methods[0]
-			position := methoda.Decl.Pos()
-			fs := info.FSet.Position(position)
-			fmt.Println(methoda.Name)
-			fmt.Println(fs.Filename, fs.Line)
-		}
-	}
-	fmt.Println("-----------------------")
 }
 
 // paths determines the paths to use.
@@ -166,78 +120,6 @@ func nodeFunc(info *godoc.PageInfo, node interface{}) string {
 	var buf bytes.Buffer
 	printer.Fprint(&buf, info.FSet, node)
 	return buf.String()
-}
-
-type DocsPackage struct {
-	Uid         string                `json:"uid"`
-	Name        string                `json:"name"`
-	IsMain      bool                  `json:"ismain"`
-	Summary     string                `json:"summary"`
-	Description string                `json:"description"`
-	ImportPath  string                `json:"importPath"`
-	Dir         string                `json:"dir"`
-	Consts      []DocsValue           `json:"consts"`
-	Types       []string              `json:"types"`
-	Vars        []DocsValue           `json:"vars"`
-	Funcs       []DocsFunc            `json:"funcs"`
-	Notes       map[string][]DocsNote `json:"notes"`
-	Examples    []DocsExample         `json:"examples"`
-	Dirs        []DocsDir             `json:"dirs"`
-}
-
-type DocsDir struct {
-	Name    string `json:"name"`
-	Path    string `json:"path"`
-	Summary string `json:"summary"`
-	HasPkg  bool   `json:"haspkg"`
-}
-
-type DocsNote struct {
-	UID         string `json:"uid"`
-	Description string `json:"description"`
-}
-
-type DocsExample struct {
-	Name string `json:"name"`
-	Code string `json:"code"`
-}
-
-type SourcePosition struct {
-	Repo   string `json:"repo"`
-	Branch string `json:"branch"`
-	File   string `json:"file"`
-	Line   int    `json:"line"`
-}
-
-type DocsValue struct {
-	Names       []string       `json:"names"`
-	Summary     string         `json:"summary"`
-	Description string         `json:"description"`
-	Code        string         `json:"code"`
-	Source      SourcePosition `json:"source"`
-}
-
-type DocsType struct {
-	Uid         string `json:"uid"`
-	Name        string `json:"name"`
-	Summary     string `json:"summary"`
-	Description string `json:"description"`
-	Code        string `json:"code"`
-
-	Consts  []DocsValue    `json:"consts"`
-	Vars    []DocsValue    `json:"vars"`
-	Funcs   []DocsFunc     `json:"funcs"`
-	Methods []DocsFunc     `json:"methods"`
-	Source  SourcePosition `json:"source"`
-}
-
-type DocsFunc struct {
-	Uid         string         `json:"uid"`
-	Name        string         `json:"name"`
-	Summary     string         `json:"summary"`
-	Description string         `json:"description"`
-	Code        string         `json:"code"`
-	Source      SourcePosition `json:"source"`
 }
 
 func ToDocfx(info *godoc.PageInfo, packageName string, packagePrefix string) (*DocsPackage, []DocsType) {
